# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AppError {
  error: string;
}

export interface ModelStats {
  totalModels: number;
  enabledModels: number;
  totalConversations: number;
  totalMessages: number;
}

export interface Model {
  id: number;
  name: string;
  modelId: string;
  /** @nullable */
  description: string | null;
  temperature: number;
  maxTokens: number;
  /** @nullable */
  systemPrompt: string | null;
  topP: number;
  enabled: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ModelInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  modelId: string;
  description?: string;
  /**
     * @minimum 0
     * @maximum 2
     */
  temperature?: number;
  /** @minimum 1 */
  maxTokens?: number;
  systemPrompt?: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  topP?: number;
  enabled?: boolean;
}

export interface ModelUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  modelId?: string;
  description?: string;
  /**
     * @minimum 0
     * @maximum 2
     */
  temperature?: number;
  /** @minimum 1 */
  maxTokens?: number;
  systemPrompt?: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  topP?: number;
  enabled?: boolean;
}

export interface Conversation {
  id: number;
  modelId: number;
  title: string;
  createdAt: string;
}

export interface ConversationInput {
  /** @minLength 1 */
  title: string;
}

export interface ChatMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  modelId: number;
  title: string;
  createdAt: string;
  messages: ChatMessage[];
}

export interface ModelMessageInput {
  content: string;
}