# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /models:
    get:
      operationId: listModels
      tags: [models]
      summary: List all configured models
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Model"
    post:
      operationId: createModel
      tags: [models]
      summary: Add a new model configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelInput"
      responses:
        "201":
          description: Created model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
  /models/{id}:
    get:
      operationId: getModel
      tags: [models]
      summary: Get a model configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
    patch:
      operationId: updateModel
      tags: [models]
      summary: Update a model configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelUpdate"
      responses:
        "200":
          description: Updated model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
    delete:
      operationId: deleteModel
      tags: [models]
      summary: Delete a model configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppError"
...
components:
  schemas:
...
    AppError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Model:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        modelId:
          type: string
        description:
          type: ["string", "null"]
        temperature:
          type: number
        maxTokens:
          type: number
        systemPrompt:
          type: ["string", "null"]
        topP:
          type: number
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - modelId
        - description
        - temperature
        - maxTokens
        - systemPrompt
        - topP
        - enabled
        - createdAt
        - updatedAt
    ModelInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        modelId:
          type: string
          minLength: 1
        description:
          type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2
        maxTokens:
          type: number
          minimum: 1
        systemPrompt:
          type: string
        topP:
          type: number
          minimum: 0
          maximum: 1
        enabled:
          type: boolean
      required:
        - name
        - modelId
    ModelUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        modelId:
          type: string
          minLength: 1
        description:
          type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2
        maxTokens:
          type: number
          minimum: 1
        systemPrompt:
          type: string
        topP:
          type: number
          minimum: 0
          maximum: 1
        enabled:
          type: boolean
...