# lib/api-zod/src/generated/types/model.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Model {
  id: number;
  name: string;
  modelId: string;
  /** @nullable */
  description: string | null;
  temperature: number;
  maxTokens: number;
  /** @nullable */
  systemPrompt: string | null;
  topP: number;
  enabled: boolean;
  createdAt: Date;
  updatedAt: Date;
}