# lib/api-zod/src/generated/types/modelInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ModelInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  modelId: string;
  description?: string;
  /**
     * @minimum 0
     * @maximum 2
     */
  temperature?: number;
  /** @minimum 1 */
  maxTokens?: number;
  systemPrompt?: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  topP?: number;
  enabled?: boolean;
}